requirement.txt
streamlit
streamlit_option_menu
google-generativeai
python-dotenv
langchain
chromadb
faiss-cpu
langchain_google_genai
PyPDF2
pptx
docx
